+++ packages/block-editor/src/components/rich-text/index.js
@@ -7,7 +7,13 @@ import { omit } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { RawHTML, useRef, useCallback, forwardRef } from '@wordpress/element';
+import {
+	RawHTML,
+	useRef,
+	useCallback,
+	forwardRef,
+	createContext,
+} from '@wordpress/element';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { children as childrenSource } from '@wordpress/blocks';
 import { useInstanceId, useMergeRefs } from '@wordpress/compose';
@@ -36,8 +42,13 @@ import { useInputRules } from './use-input-rules';
 import { useEnter } from './use-enter';
 import { useFormatTypes } from './use-format-types';
 import { useRemoveBrowserShortcuts } from './use-remove-browser-shortcuts';
+import { useShortcuts } from './use-shortcuts';
+import { useInputEvents } from './use-input-events';
 import FormatEdit from './format-edit';
 import { getMultilineTag, getAllowedFormats } from './utils';
+
+export const keyboardShortcutContext = createContext();
+export const inputEventContext = createContext();
 
 /**
  * Removes props used for the native version of RichText so that they are not
@@ -213,7 +224,7 @@ function RichTextWrapper(
 		);
 	}
 
-	const { value, onChange, onFocus, ref: richTextRef } = useRichText( {
+	const { value, onChange, ref: richTextRef } = useRichText( {
 		value: adjustedValue,
 		onChange( html, { __unstableFormats, __unstableText } ) {
 			adjustedOnChange( html );
@@ -243,6 +254,9 @@ function RichTextWrapper(
 
 	useCaretInFormat( { value } );
 	useMarkPersistent( { html: adjustedValue, value } );
+
+	const keyboardShortcuts = useRef( new Set() );
+	const inputEvents = useRef( new Set() );
 
 	function onKeyDown( event ) {
 		const { keyCode } = event;
@@ -283,20 +297,26 @@ function RichTextWrapper(
 		}
 	}
 
+	function onFocus() {
+		anchorRef.current.focus();
+	}
+
 	const TagName = tagName;
 	const content = (
 		<>
-			{ isSelected &&
-				children &&
-				children( { value, onChange, onFocus } ) }
 			{ isSelected && (
-				<FormatEdit
-					value={ value }
-					onChange={ onChange }
-					onFocus={ onFocus }
-					formatTypes={ formatTypes }
-					forwardedRef={ anchorRef }
-				/>
+				<keyboardShortcutContext.Provider value={ keyboardShortcuts }>
+					<inputEventContext.Provider value={ inputEvents }>
+						{ children && children( { value, onChange, onFocus } ) }
+						<FormatEdit
+							value={ value }
+							onChange={ onChange }
+							onFocus={ onFocus }
+							formatTypes={ formatTypes }
+							forwardedRef={ anchorRef }
+						/>
+					</inputEventContext.Provider>
+				</keyboardShortcutContext.Provider>
 			) }
 			{ isSelected && hasFormats && (
 				<FormatToolbarContainer
@@ -323,6 +343,8 @@ function RichTextWrapper(
 						onReplace,
 					} ),
 					useRemoveBrowserShortcuts(),
+					useShortcuts( keyboardShortcuts ),
+					useInputEvents( inputEvents ),
 					useUndoAutomaticChange(),
 					usePasteHandler( {
 						isSelected,