+++ packages/edit-post/src/store/selectors.js
@@ -189,9 +189,14 @@ export function isModalActive( state, modalName ) {
  *
  * @return {boolean} Is active.
  */
-export function isFeatureActive( state, feature ) {
-	return get( state.preferences.features, [ feature ], false );
-}
+export const isFeatureActive = createRegistrySelector(
+	( select ) => ( state, feature ) => {
+		return select( interfaceStore ).isFeatureActive(
+			'core/edit-post',
+			feature
+		);
+	}
+);
 
 /**
  * Returns true if the plugin item is pinned to the header.
@@ -360,6 +365,17 @@ export function isListViewOpened( state ) {
  */
 export function isEditingTemplate( state ) {
 	return state.isEditingTemplate;
+}
+
+/**
+ * Returns true if meta boxes are initialized.
+ *
+ * @param {Object} state Global application state.
+ *
+ * @return {boolean} Whether meta boxes are initialized.
+ */
+export function areMetaBoxesInitialized( state ) {
+	return state.metaBoxes.initialized;
 }
 
 /**