+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,12 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [**] Enable embed preview for a list of providers (for now only YouTube and Twitter) [#34446]
+## 1.60.1
+-   [*] RNmobile: Fix the cancel button on Block Variation Picker / Columns Block. [#34249]
+-   [*] Column block: Fix Android close button alignment. [#34332]
+
+## 1.60.0
 -   [**] Embed block: Add "Resize for smaller devices" setting. [#33654]
 
 ## 1.59.2